Task 3 - done
Program the Money class (a class object that operates with one currency) to work with money.
The class should have a field for storing a whole part of money (dollars, euros, hryvnias, etc.) and a field for storing kopecks (cents, eurocents, kopecks, etc.).
Implement methods for displaying sums on the screen and setting values for parts.
Based on the Money class, create a Product class to work with a product or product. Implement a method that allows you to reduce the price by a given number.
For each of the classes, implement the necessary methods and fields.